Add function to get deployment addresses from settings
